Integrate MCPE 1.0 features: assets, UI, hunger system, infinite world

+++ web-build/src/world/level/storage/ExternalFileLevelStorage.h
@@ -19,7 +19,7 @@ class RegionFile;
 
 typedef struct UnsavedLevelChunk
 {
-	int				pos;
+	long long		pos;
 	RakNet::TimeMS	addedToList;
 	LevelChunk*		chunk;
 } UnsavedLevelChunk;
